packages/morag-document/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morag-document"
version = "0.1.0"
description = "Document processing service for MoRAG (Modular Retrieval Augmented Generation)"
authors = [
    {name = "MoRAG Team"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "morag-core>=0.1.0",
    "morag-embedding>=0.1.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
    "openpyxl>=3.0.10",
    "python-pptx>=0.6.21",
    "beautifulsoup4>=4.11.0",
    "markdown>=3.4.0",
    "nltk>=3.7",
    "spacy>=3.4.0",
    "langdetect>=1.0.9",
    "aiohttp>=3.8.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/morag"
Documentation = "https://github.com/yourusername/morag#readme"
Repository = "https://github.com/yourusername/morag.git"
Issues = "https://github.com/yourusername/morag/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true